DOC: More doc fixes to pass tests
This does *not* fix everything, I am hoping for an automated tester
based on doctestplus for this (I actually have that mostly ready)

Doing this here woul dbe quite spammy though!

+++ numpy/ma/core.py
@@ -7564,7 +7564,7 @@ def diff(a, /, n=1, axis=-1, prepend=np._NoValue, append=np._NoValue):
     >>> np.ma.diff(u8_arr)
     masked_array(data=[255],
                  mask=False,
-           fill_value=999999,
+           fill_value=np.int64(999999),
                 dtype=uint8)
     >>> u8_arr[1,...] - u8_arr[0,...]
     255
@@ -7576,7 +7576,7 @@ def diff(a, /, n=1, axis=-1, prepend=np._NoValue, append=np._NoValue):
     >>> np.ma.diff(i16_arr)
     masked_array(data=[-1],
                  mask=False,
-           fill_value=999999,
+           fill_value=np.int64(999999),
                 dtype=int16)
 
     Examples
